module2-task1. Project with mockies
