+++ src/view/list-view.js
@@ -1,0 +1,8 @@
+export const createListTemplate = () => (
+  `<section class="films">
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+        <div class="films-list__container"></div>
+      </section>
+    </section>`
+);
