--- src/view/number-of-movies-view.js
@@ -1,4 +1,0 @@
-export const createNumberOfMovies = () => (
-  `<p>130 291 movies inside
-  </p>`
-);
