--- src/view/profile-rating-view.js
@@ -1,6 +1,0 @@
-export const createProfileRating = () => (
-  `<section class="header__profile profile">
-<p class="profile__rating">Movie buff</p>
-<img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-</section>`
-);
