+++ src/view/show-more-button-view.js
@@ -1,4 +1,4 @@
-export const createShowMoreButton = () => (
+export const createShowMoreButtonTemplate = () => (
   `<button class="films-list__show-more">Show more
   </button>`
 );
