+++ src/view/stats-view.js
